[SW-211][deleted]checkCollision 함수 더이상 사용하지 않아서 제거
